refactor: Refactor mock handlers structure
