+++ app/mocks/handlers/index.ts
@@ -1,0 +1,4 @@
+import { takenLectureHandlers } from './taken-lecture-handler.mock';
+import { userHandlers } from './user-handler.mock';
+
+export const handlers = [...userHandlers, ...takenLectureHandlers];
